iPad support: use popovers instead of modal view.
Apple requires that image pickers are displayed using UIPopoverController when on an iPad. This implements that. If the options popoverRectX, popoverRectY, popoverRectWidth or popoverRectHeight are passed, these are used as the location of the popover's arrow. It's a good idea to pass the offset and size of a clicked element, or the location of a click event (with a small size).